+++ source/migration/data/Version20201029110624.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OxidEsales\EshopCommunity\Migrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+final class Version20201029110624 extends AbstractMigration
+{
+    public function up(Schema $schema): void
+    {
+        $this->connection->getDatabasePlatform()->registerDoctrineTypeMapping('enum', 'string');
+
+        $table = $schema->getTable('oxuser');
+        $table->addIndex(['oxrights'], 'OXRIGHTS');
+    }
+
+    public function down(Schema $schema): void
+    {
+        $this->connection->getDatabasePlatform()->registerDoctrineTypeMapping('enum', 'string');
+
+        $table = $schema->getTable('oxuser');
+        $table->dropIndex('oxrights');
+    }
+}